Crate arcella_wasmtime

Wasmtime-specific utilities for Arcella.

This crate provides conversion from wasmtime’s internal component model types into the stable, serializable types defined in arcella-types.

It is intended for use by the Arcella runtime and CLI tools that need to inspect WebAssembly components using Wasmtime as the engine.

Re-exports

pub use error::ArcellaWasmtimeError;

pub use error::Result;

pub use manifest::ComponentManifestExt;

Modules

error

manifest

Traits

ComponentItemSpecExt

Extension trait to convert wasmtime::component::types::ComponentItem into ComponentItemSpec.

ComponentTypeExt

Extension trait for wasmtime::component::types::Component.